requirements.txt

Pillow
streamlit
opencv-python
